refactor: move skeleton & error-indicator out of record-selection
