+++ src/components/organisms/RecordSelection/index.tsx
@@ -2,10 +2,8 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { styled } from 'styled-components';
 
-import { Skeleton } from '^/components/atoms/Skeleton';
 import { RecordListCard } from '^/components/molecules/RecordListCard';
 import { ShmupRecordPreview } from '^/types';
-import { ErrorIndicator } from '^/components/molecules/ErrorIndicator';
 
 const Root = styled.div`
   display: flex;
@@ -28,43 +26,12 @@ const RecordSelectionLink = styled(Link)`
   text-decoration: none;
 `;
 
-const RecordListCardSkeletonRoot = styled.div`
-  display: flex;
-  flex-direction: column;
-  row-gap: 8px;
-`;
-
-function RecordListCardSkeleton() {
-  return (
-    <RecordListCardSkeletonRoot>
-      <Skeleton width="300px" height="200px" borderRadius="16px" />
-      <Skeleton width="300px" height="20px" borderRadius="10px" />
-    </RecordListCardSkeletonRoot>
-  );
-}
-
 interface Props {
   recordPreviews: ShmupRecordPreview[];
-  isLoading: boolean;
-  isError: boolean;
 }
 
-export function RecordSelection({ recordPreviews, isLoading, isError }: Props) {
+export function RecordSelection({ recordPreviews }: Props) {
   const renderRecordSelectionArea = (() => {
-    if (isLoading) {
-      return (
-        <>
-          <RecordListCardSkeleton />
-          <RecordListCardSkeleton />
-          <RecordListCardSkeleton />
-        </>
-      );
-    }
-
-    if (isError) {
-      return <ErrorIndicator title="목록을 불러오는 중 오류가 발생했습니다." />;
-    }
-
     if (recordPreviews.length === 0) {
       return '현재 등록된 기록이 없습니다.';
     }
