+++ src/pages/RecordListPage/index.tsx
@@ -5,6 +5,8 @@ import styled from 'styled-components';
 import { textsForArticle } from '^/constants/texts';
 import { useShmupRecordPreviewList } from '^/hooks/useShmupRecordPreviewList';
 import { RecordSelection } from '^/components/organisms/RecordSelection';
+import { ErrorIndicator } from '^/components/molecules/ErrorIndicator';
+import { Skeleton } from '^/components/atoms/Skeleton';
 
 const Root = styled.div`
   padding-left: 15px;
@@ -22,6 +24,28 @@ const Title = styled.h1`
   font-weight: 700;
 `;
 
+const RecordListCardSkeletonListWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 16px;
+`;
+
+const RecordListCardSkeletonRoot = styled.div`
+  display: flex;
+  flex-direction: column;
+  row-gap: 8px;
+`;
+
+function RecordListCardSkeleton() {
+  return (
+    <RecordListCardSkeletonRoot>
+      <Skeleton width="300px" height="200px" borderRadius="16px" />
+      <Skeleton width="300px" height="20px" borderRadius="10px" />
+    </RecordListCardSkeletonRoot>
+  );
+}
+
 export function RecordListPage() {
   const { typeId } = useParams();
 
@@ -29,10 +53,36 @@ export function RecordListPage() {
     return null;
   }
 
+  const { recordPreviews, isLoading, isError } =
+    useShmupRecordPreviewList(typeId);
+
+  const renderRecordSelectionArea = (() => {
+    if (isLoading) {
+      return (
+        <RecordListCardSkeletonListWrapper>
+          <RecordListCardSkeleton />
+          <RecordListCardSkeleton />
+          <RecordListCardSkeleton />
+          <RecordListCardSkeleton />
+        </RecordListCardSkeletonListWrapper>
+      );
+    }
+
+    if (isError) {
+      return <ErrorIndicator title="목록을 불러오는 중 오류가 발생했습니다." />;
+    }
+
+    if (recordPreviews.length === 0) {
+      return '현재 등록된 기록이 없습니다.';
+    }
+
+    return <RecordSelection recordPreviews={recordPreviews} />;
+  })();
+
   return (
     <Root>
       <Title>{textsForArticle[typeId] ?? typeId} 기록 목록</Title>
-      <RecordSelection {...useShmupRecordPreviewList(typeId)} />
+      {renderRecordSelectionArea}
     </Root>
   );
 }
